requirements.txt
google-genai
dotenv
ipython